requirements.txt
playwright
psycopg2-binary
faker
python-dotenv
pandas
